requirements.txt
Django==2.1.10
beautifulsoup4==4.7.1
requests==2.21.0
PyMySQL==0.9.3
django-phonenumber-field==2.3.1
phonenumbers==8.10.9
google-cloud-kms==1.1.0
